# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ConstellationEntry": {
        "additionalProperties": false,
        "description": "One constellation of `GET /catalogs/constellations` (brief l.148-150).",
        "properties": {
          "abbr": {
            "examples": [
              "Ori"
            ],
            "maxLength": 3,
            "minLength": 3,
            "title": "Abbr",
            "type": "string"
          },
          "boundary": {
            "description": "IAU boundary polygon, ICRS J2000 `[ra_deg, dec_deg]` vertices, closed ring.",
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array"
            },
            "title": "Boundary",
            "type": "array"
          },
          "boundary_parts": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "maxItems": 2,
                    "minItems": 2,
                    "prefixItems": [
                      {
                        "type": "number"
                      },
                      {
                        "type": "number"
                      }
                    ],
                    "type": "array"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Every boundary polygon when the constellation has more than one (Serpens); `boundary` then holds the first part. Absent otherwise.",
            "title": "Boundary Parts"
          },
          "genitive": {
            "examples": [
              "Orionis"
            ],
            "title": "Genitive",
            "type": "string"
          },
          "label": {
            "$ref": "#/components/schemas/ConstellationLabel",
            "description": "Suggested label position."
          },
          "latin": {
            "examples": [
              "Orion"
            ],
            "title": "Latin",
            "type": "string"
          },
          "lines": {
            "description": "Line segments as pairs of Hipparcos identifiers (Stellarium `modern`).",
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "integer"
                },
                {
                  "type": "integer"
                }
              ],
              "type": "array"
            },
            "title": "Lines",
            "type": "array"
          }
        },
        "required": [
          "abbr",
          "latin",
          "genitive",
          "lines",
          "boundary",
          "label"
        ],
        "title": "ConstellationEntry",
        "type": "object"
      },
      "ConstellationLabel": {
        "additionalProperties": false,
        "properties": {
          "dec_deg": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Dec Deg",
            "type": "number"
          },
          "ra_deg": {
            "exclusiveMaximum": 360.0,
            "minimum": 0.0,
            "title": "Ra Deg",
            "type": "number"
          }
        },
        "required": [
          "ra_deg",
          "dec_deg"
        ],
        "title": "ConstellationLabel",
        "type": "object"
      },
...
      "ConstellationsResponse": {
        "additionalProperties": false,
        "description": "Body of `GET /catalogs/constellations`.",
        "properties": {
          "constellations": {
            "items": {
              "$ref": "#/components/schemas/ConstellationEntry"
            },
            "title": "Constellations",
            "type": "array"
          },
          "culture": {
            "const": "modern",
            "default": "modern",
            "title": "Culture",
            "type": "string"
          }
        },
        "required": [
          "constellations"
        ],
        "title": "ConstellationsResponse",
        "type": "object"
      },
...
      "DsoEntry": {
        "additionalProperties": false,
        "description": "One row of `GET /catalogs/dso` (brief l.144-146): an OpenNGC object, J2000.",
        "properties": {
          "con": {
            "description": "IAU constellation abbreviation.",
            "maxLength": 3,
            "minLength": 3,
            "title": "Con",
            "type": "string"
          },
          "dec_deg": {
            "description": "ICRS declination, degrees.",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Dec Deg",
            "type": "number"
          },
          "id": {
            "description": "Canonical OpenNGC id: catalog prefix plus number without zero padding, spaces replaced by `_` (`NGC224`, `IC434`, `Mel22`).",
            "examples": [
              "NGC7000"
            ],
            "title": "Id",
            "type": "string"
          },
          "mag": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "V magnitude, else B magnitude.",
            "title": "Mag"
          },
          "major_arcmin": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Major axis, arcminutes.",
            "title": "Major Arcmin"
          },
          "messier": {
            "anyOf": [
              {
                "maximum": 110.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Messier number.",
            "examples": [
              31
            ],
            "title": "Messier"
          },
          "minor_arcmin": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Minor axis, arcminutes.",
            "title": "Minor Arcmin"
          },
          "names": {
            "description": "Common names from OpenNGC.",
            "examples": [
              [
                "North America Nebula"
              ]
            ],
            "items": {
              "type": "string"
            },
            "title": "Names",
            "type": "array"
          },
          "pa_deg": {
            "anyOf": [
              {
                "exclusiveMaximum": 360.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Major-axis position angle, north through east.",
            "title": "Pa Deg"
          },
          "ra_deg": {
            "description": "ICRS right ascension, degrees.",
            "exclusiveMaximum": 360.0,
            "minimum": 0.0,
            "title": "Ra Deg",
            "type": "number"
          },
          "type": {
            "description": "Object class mapped from the OpenNGC type code.",
            "enum": [
              "galaxy",
              "open_cluster",
              "globular_cluster",
              "planetary_nebula",
              "nebula",
              "other"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "names",
          "type",
          "ra_deg",
          "dec_deg",
          "con"
        ],
        "title": "DsoEntry",
        "type": "object"
      },
...
      "Problem": {
        "additionalProperties": false,
        "description": "RFC 9457 problem document (`type`, `title`, `status`, `detail`, `instance`, `errors[]`).",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Explanation specific to this occurrence.",
            "examples": [
              "unknown body ids: vulcan"
            ],
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ProblemError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Per-parameter validation errors (400 only).",
            "title": "Errors"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Path of the request (never its query string).",
            "examples": [
              "/api/v1/sky/frame"
            ],
            "title": "Instance"
          },
          "range_tt": {
            "anyOf": [
              {
                "maxItems": 2,
                "minItems": 2,
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Valid TT Julian Date range when the request fell outside data coverage (422).",
            "examples": [
              [
                2396758.5,
                2506000.5
              ]
            ],
            "title": "Range Tt"
          },
          "status": {
            "description": "HTTP status code.",
            "examples": [
              400
            ],
            "maximum": 599,
            "minimum": 400,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short summary of the problem type.",
            "examples": [
              "Invalid parameter"
            ],
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "URI identifying the problem type; every type is documented in docs/api.md.",
            "examples": [
              "https://github.com/j-about/InterSidera/blob/master/docs/api.md#problem-invalid-parameter"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ],
        "title": "Problem",
        "type": "object"
      },
      "ProblemError": {
        "additionalProperties": false,
        "description": "One invalid input, in pydantic's `loc`/`msg`/`type` vocabulary.",
        "properties": {
          "loc": {
            "description": "Location of the offending value: `query` and the parameter name.",
            "examples": [
              [
                "query",
                "n"
              ]
            ],
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Loc",
            "type": "array"
          },
          "msg": {
            "description": "Human-readable explanation.",
            "examples": [
              "Input should be less than or equal to 64"
            ],
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "description": "Machine-readable error type (pydantic's error type).",
            "examples": [
              "less_than_equal"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ProblemError",
        "type": "object"
      },
...
      "StarIndexEntry": {
        "additionalProperties": false,
        "description": "One row of `GET /catalogs/stars/index` (brief l.141-142).",
        "properties": {
          "con": {
            "description": "IAU constellation abbreviation.",
            "examples": [
              "Ori"
            ],
            "maxLength": 3,
            "minLength": 3,
            "title": "Con",
            "type": "string"
          },
          "hip": {
            "description": "Hipparcos identifier.",
            "examples": [
              27989
            ],
            "minimum": 1.0,
            "title": "Hip",
            "type": "integer"
          },
          "names": {
            "$ref": "#/components/schemas/StarNames"
          }
        },
        "required": [
          "hip",
          "names",
          "con"
        ],
        "title": "StarIndexEntry",
        "type": "object"
      },
      "StarNames": {
        "additionalProperties": false,
        "description": "Names of a designated star; every field is optional but at least one is present.",
        "properties": {
          "bayer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Bayer designation: Greek letter, optional superscript, IAU abbreviation.",
            "examples": [
              "α Ori"
            ],
            "title": "Bayer"
          },
          "flamsteed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Flamsteed designation: number and IAU abbreviation.",
            "examples": [
              "58 Ori"
            ],
            "title": "Flamsteed"
          },
          "proper": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "IAU proper name (HYG `proper`), never translated.",
            "examples": [
              "Betelgeuse"
            ],
            "title": "Proper"
          }
        },
        "title": "StarNames",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Stateless, cacheable GET API behind InterSidera: Skyfield is the astronomical authority, clients only rotate, interpolate and apply the corrections it supplies. Angles in degrees, distances in au, times as TT Julian Dates unless suffixed `_utc`. Errors are RFC 9457 problem documents (`application/problem+json`).",
    "title": "InterSidera Sky API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/catalogs/constellations": {
      "get": {
        "description": "The 88 IAU constellations: Stellarium `modern` line segments as pairs of Hipparcos identifiers (so lines follow the rendered stars), IAU boundary polygons as closed rings of ICRS J2000 `[ra_deg, dec_deg]` vertices (`boundary_parts` lists every ring when a constellation has more than one, Serpens), Latin and genitive names and a label position. Answers 503 while the Stellarium or d3-celestial data is missing (`/health` lists `constellations` in `missing`). Served with `Cache-Control: public, max-age=3600, stale-while-revalidate=86400` and a strong `ETag` (the artifact SHA-256, bare in `/meta.catalogs`); `If-None-Match` -> 304.",
        "operationId": "get_constellations_api_v1_catalogs_constellations_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConstellationsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "Not modified (`If-None-Match` matched the ETag); `ETag` and `Cache-Control` are repeated, the body is empty."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Data not ready (starting, or degraded for this endpoint); retry after `Retry-After`."
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Any other error, as an RFC 9457 problem document."
          }
        },
        "summary": "Constellation lines, boundaries and labels (modern culture)",
        "tags": [
          "catalogs"
        ]
      }
    },
    "/api/v1/catalogs/dso": {
      "get": {
        "description": "Every Messier object plus the NGC/IC objects with `mag <= 14` or a major axis of at least 5 arcminutes, J2000 ICRS. `id` is the OpenNGC name without zero padding and with spaces replaced by `_` (`NGC224`, `IC434`, `Mel22`); `messier` carries the Messier number when the object has one; `type` is the OpenNGC class mapped to six values. Answers 503 while the OpenNGC data is missing (`/health` lists `dso` in `missing`). Served with `Cache-Control: public, max-age=3600, stale-while-revalidate=86400` and a strong `ETag` (the artifact SHA-256, bare in `/meta.catalogs`); `If-None-Match` -> 304.",
        "operationId": "get_dso_api_v1_catalogs_dso_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DsoEntry"
                  },
                  "title": "Response Get Dso Api V1 Catalogs Dso Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "Not modified (`If-None-Match` matched the ETag); `ETag` and `Cache-Control` are repeated, the body is empty."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Data not ready (starting, or degraded for this endpoint); retry after `Retry-After`."
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Any other error, as an RFC 9457 problem document."
          }
        },
        "summary": "Deep-sky objects (OpenNGC subset)",
        "tags": [
          "catalogs"
        ]
      }
    },
    "/api/v1/catalogs/stars": {
      "get": {
        "description": "Every Hipparcos star with a position, as the little-endian binary format `SKYS` v1 (`application/octet-stream`). Header, 24 bytes: magic `SKYS` (4 bytes), `u32` version = 1, `u32` count `n`, `f64` epoch_tt (2451545.0 = J2000; Hipparcos positions are propagated from J1991.25 by Skyfield at build time), `u32` flags (reserved, 0). Then the columns, packed without padding: `dir` f32[3n] ICRF unit vector at the epoch, `pm` f32[3n] tangential proper-motion velocity in ICRF in radians per Julian year (parallax and radial velocity ignored), `mag` i16[n] Johnson V in millimagnitudes, `bv` i16[n] B-V in millimagnitudes (32767 when unknown), `hip` u32[n] Hipparcos identifier; total 24 + 32 n bytes. Stars are sorted by magnitude ascending, so any prefix is a valid brighter-than subset. Shader rule: `dir(t) = normalize(dir + pm * years_since_epoch)`, then aberration `normalize(dir(t) + observer_velocity / c)`. Served with `Cache-Control: public, max-age=3600, stale-while-revalidate=86400` and a strong `ETag` (the artifact SHA-256, bare in `/meta.catalogs`); `If-None-Match` -> 304.",
        "operationId": "get_stars_api_v1_catalogs_stars_get",
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {}
            },
            "description": "The SKYS v1 file (`24 + 32 n` bytes, `n` in the header)."
          },
          "304": {
            "description": "Not modified (`If-None-Match` matched the ETag); `ETag` and `Cache-Control` are repeated, the body is empty."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Data not ready (starting, or degraded for this endpoint); retry after `Retry-After`."
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Any other error, as an RFC 9457 problem document."
          }
        },
        "summary": "Star catalog, binary SKYS v1",
        "tags": [
          "catalogs"
        ]
      }
    },
    "/api/v1/catalogs/stars/index": {
      "get": {
        "description": "The stars that carry an IAU proper name, a Bayer or a Flamsteed designation (about 3 400 of the Hipparcos catalog), with their constellation. Proper names are the IAU names and are not translated; `bayer` is the Greek letter with an optional superscript and the IAU abbreviation (`α Ori`), `flamsteed` the number and abbreviation (`58 Ori`). Served with `Cache-Control: public, max-age=3600, stale-while-revalidate=86400` and a strong `ETag` (the artifact SHA-256, bare in `/meta.catalogs`); `If-None-Match` -> 304.",
        "operationId": "get_stars_index_api_v1_catalogs_stars_index_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StarIndexEntry"
                  },
                  "title": "Response Get Stars Index Api V1 Catalogs Stars Index Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "304": {
            "description": "Not modified (`If-None-Match` matched the ETag); `ETag` and `Cache-Control` are repeated, the body is empty."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Data not ready (starting, or degraded for this endpoint); retry after `Retry-After`."
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Any other error, as an RFC 9457 problem document."
          }
        },
        "summary": "Named or designated stars for search and labels",
        "tags": [
          "catalogs"
        ]
      }
    },
...
  },
...
}